// src/app/services/models/user-notification-preferences.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserNotificationPreferences {
  createdBy?: string;
  createdDate?: string;
  id?: number;
  lastModifiedBy?: string;
  lastModifiedDate?: string;
  notifyOnCancelledTransfers?: boolean;
  notifyOnDeposit?: boolean;
  notifyOnExecutedTransfers?: boolean;
  notifyOnPaymentFailure?: boolean;
  notifyOnReceive?: boolean;
  notifyOnScheduledTransfers?: boolean;
  notifyOnSend?: boolean;
  notifyOnWithdraw?: boolean;
}